requirements.txt
Django==2.2
django-mdeditor==0.1.13
Jinja2==2.10
Markdown==3.1
MarkupSafe==1.1.1
Pillow==6.0.0
psycopg2==2.7.7
pytz==2018.9
sqlparse==0.3.0
